web/requirements.txt
alembic
Flask
gevent
hiredis
redis==2.10.6  # https://github.com/celery/celery/issues/5175
PyTumblr
psycopg2
WebHelpers
flower
celery
SQLAlchemy
gunicorn
classtools
beautifulsoup4
html5lib
